groups sorting/bugfix/menu changes 1. missing local cleanup on db group remove. 2. PlannerGoals: all element calculcations into memo/callback. Added "Sort empty groups to bottom" settings/func. 3. use menu component+hook 4. moved Allow all goals > goals menu

+++ src/util/hooks/useGoals.ts
@@ -105,10 +105,10 @@ function useGoals() {
   }, [_setGoals]);
 
   const removeAllGoalsFromGroup = useCallback(
-    async (groupName: string) => {
+    async (groupName: string, cleanLocal?: boolean) => {
       const filteredArray = goals.filter((goal) => goal.group_name !== groupName);
       _setGoals(filteredArray);
-
+      if (cleanLocal) return;
       const { error } = await supabase.from("goals").delete().eq("group_name", groupName);
       handlePostgrestError(error);
     },
